+++ jwst/extract_2d/tests/test_grisms.py
@@ -160,6 +160,40 @@ def get_reference_files(datamodel):
     return refs
 
 
+@pytest.fixture(params=tsgrism_filters)
+def tsgrism_inputs(request):
+    def _add_missing_key(missing_key=None):
+        tso_kw = wcs_tso_kw.copy()
+        tso_kw.update({'xref_sci': 887.0, 'yref_sci': 35.0})
+
+        if missing_key is not None:
+            tso_kw[missing_key] = None
+
+        hdu = create_hdul(
+            exptype='NRC_TSGRISM',
+            pupil='GRISMR',
+            filtername=request.param,
+            detector='NRCALONG',
+            subarray='SUBGRISM256',
+            wcskeys=tso_kw,
+            channel='LONG',
+            module='A',
+        )
+
+        image_model = CubeModel(hdu)
+
+        return image_model, get_reference_files(image_model)
+
+    return _add_missing_key
+
+
+@pytest.mark.parametrize('key', ['xref_sci', 'yref_sci'])
+def test_extract_tso_object_fails_without_xref_yref(tsgrism_inputs, key):
+    with pytest.raises(ValueError):
+        image_model, refs = tsgrism_inputs(missing_key=key)
+        extract_tso_object(image_model, reference_files=refs)
+
+
 @pytest.mark.filterwarnings("ignore: Card is too long")
 def test_create_box_fits():
     """Make sure that a box is created around a source catalog object.